+++ src/pages/HomePage/HomePage.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import SearchBar from './SearchBar';
+import SearchBar from '../../components/SearchBar';
 
 class HomePage extends React.Component {
   render() {
